feat : 상품 리뷰 등록 (#21)
- @Transactional 어노테이션 수정
- 테스트를 위해 memberId 를 임시적으로 @RequestParam 으로 수정
- 상품 상태 변경에 거래완료 등록 로직 추가
- 리뷰 등록 API 추가

+++ src/main/java/com/cucumber/market/api/service/item/ItemService.java
@@ -13,11 +13,11 @@ import java.util.Map;
 @Service
 @RequiredArgsConstructor
 @Slf4j
+@Transactional
 public class ItemService {
 
     private final ItemMapper itemMapper;
 
-    @Transactional
     public Map addItem(Long memberId, ItemDto.addItemDto itemDto) {
         LinkedHashMap<String, Object> result = new LinkedHashMap<>();
 
@@ -29,7 +29,8 @@ public class ItemService {
         return result;
     }
 
-    @Transactional
+
+    @Transactional(readOnly = true)
     public Map getItem(Long itemId) {
         LinkedHashMap<String, Object> result = new LinkedHashMap<>();
 
@@ -39,30 +40,44 @@ public class ItemService {
         return result;
     }
 
-    @Transactional
+
     public Map modifyItem(Long memberId, Long itemId, ItemDto.modifyItemDto itemDto) {
         LinkedHashMap<String, Object> result = new LinkedHashMap<>();
 
-        //userMapper.findById(memberId); -> 상품 수정 권한 확인
+        Map item = itemMapper.selectItem(itemId).orElseThrow(IllegalArgumentException::new);
+        if (item.get("memberId").equals(memberId.intValue())) {  //상품 수정 권한 확인
 
-        itemMapper.updateItem(itemId, itemDto);
-        result.put("item", itemDto);
+            itemMapper.updateItem(itemId, itemDto);
+            result.put("item", itemDto);
+        }
 
         return result;
     }
 
-    public Map modifyItemStatus(Long memberId, Long itemId, ItemStatus itemStatus) {
+
+    public Map modifyItemStatus(Long memberId, Long itemId, ItemDto.modifyItemStatusDto itemDto) {
         LinkedHashMap<String, Object> result = new LinkedHashMap<>();
+        ItemStatus itemStatus = itemDto.getItemStatus();
 
-        //userMapper.findById(memberId); -> 상품 상태 수정 권한 확인
+        Map item = itemMapper.selectItem(itemId).orElseThrow(IllegalArgumentException::new);
+        if (item.get("memberId").equals(memberId.intValue())) {  //상품 상태 수정 권한 확인
 
-        itemMapper.updateItemStatus(itemId, itemStatus);
-        result.put("itemId", itemId);
-        result.put("itemStatus", itemStatus);
+            itemMapper.deleteBuyerReview(itemId);  //기존 구매자 리뷰 삭제
 
-        return  result;
+            if (itemStatus == ItemStatus.CLOSED) {  //거래 완료 시 구매자 후기 테이블에도 등록
+                itemMapper.insertBuyerReview(itemId, itemDto.getClientId(), null);
+            }
+
+            itemMapper.updateItemStatus(itemId, itemStatus);
+            result.put("itemId", itemId);
+            result.put("itemStatus", itemStatus);
+        }
+
+        return result;
     }
 
+
+    @Transactional(readOnly = true)
     public Map getItems(String itemName, ItemStatus itemStatus) {
         LinkedHashMap<String, Object> result = new LinkedHashMap<>();
 
@@ -71,21 +86,25 @@ public class ItemService {
         return result;
     }
 
+
     public Map deleteItem(Long memberId, Long itemId) {
         LinkedHashMap<String, Object> result = new LinkedHashMap<>();
 
-        //userMapper.findById(memberId); -> 상품 삭제 권한 확인
+        Map item = itemMapper.selectItem(itemId).orElseThrow(IllegalArgumentException::new);
+        if (item.get("memberId").equals(memberId.intValue())) {  //상품 삭제 권한 확인
 
-        itemMapper.deleteItem(itemId);
-        result.put("itemId", itemId);
+            itemMapper.deleteItem(itemId);
+            result.put("itemId", itemId);
+        }
 
         return result;
     }
 
+
     public Map addLike(Long itemId, Long memberId) {
         LinkedHashMap<String, Object> result = new LinkedHashMap<>();
 
-        //userMapper.findById(memberId); -> 상품 삭제 권한 확인
+        //userMapper.findById(memberId); -> 상품 좋아요 등록 권한 확인
 
         itemMapper.insertLike(itemId, memberId);
         result.put("itemId", itemId);
@@ -93,14 +112,39 @@ public class ItemService {
         return result;
     }
 
+
     public Map deleteLike(Long itemId, Long memberId) {
         LinkedHashMap<String, Object> result = new LinkedHashMap<>();
 
-        //userMapper.findById(memberId); -> 상품 삭제 권한 확인
+        //userMapper.findById(memberId); -> 상품 좋아요 삭제 권한 확인
 
         itemMapper.deleteLike(itemId, memberId);
         result.put("itemId", itemId);
 
         return result;
     }
- }
+
+
+    public Map addReview(Long itemId, Long memberId, ItemDto.reviewDto reviewDto) {
+        LinkedHashMap<String, Object> result = new LinkedHashMap<>();
+
+        Map item = itemMapper.selectItem(itemId).orElseThrow(IllegalArgumentException::new);
+        if (item.get("memberId").equals(memberId.intValue())) {  //판매자 여부 판별
+
+            itemMapper.insertSellerReview(itemId, reviewDto);
+        } else {
+            Map buyerReview = itemMapper.selectBuyerReview(itemId).orElseThrow(IllegalArgumentException::new);
+
+            if (!buyerReview.get("memberId").equals(memberId.intValue())) {   //구매자 여부 판별
+                //throw Exception
+            }
+            itemMapper.updateBuyerReview(itemId, reviewDto);
+
+            result.put("itemId", itemId);
+            result.put("review", reviewDto.getReview());
+        }
+
+        return result;
+    }
+}
+
